core: split audit header helpers
